[skif] Avoid resolving a simpled deferred color filter
When removing SK_RESOLVE_FILTERS_BEFORE_RESTORE, there were two failures
in Flutters tests. One was a rounding issue because a color filter now
was operating on the original values instead of the 8-bit values. The
other involved applying a "group opacity" optimization to a drawImage
call that also had an image filter and expecting that the
alpha-modulated output match the manually modulated original image.
This boils down to adding an alpha value on the SkPaint.

The Flutter test in question used a non-AA paint with nearest neighbor
filtering and a dst rect aligned on 1/2px boundaries. With the legacy
build flag enabled, both the opaque and the "group opacity" case would
draw the non-AA image to a new layer (possibly with alpha), and then
pass that to an erode/dilate filter. The non-AA rendering meant that
there were no pixels with partial coverage.

With the build flag removed, both cases skipped the auto layer within
SkCanvas. The opaque case deferred the src->dst transform and nearest
neighbor filtering into the erode/dilate evaluation. The use of
nearest neighbor filtering produced an equivalent result to the
original's non-AA snapping in the layer itself.

However, for the "group opacity" case, the alpha is now applied as
a color filter to the FilterResult before going to the erode/dilate
filter. This is deferred but because erode/dilate pass in the
kSampledRepeatedly flag, asShader() resolves the image. FilterResult
uses AA by default because it gives more stable behavior when images
are being processed by later filters, but in this case we now
introduce partial coverage on the border pixels because of the
half-pixel aligned dst rect in the deferred transform.

While this actually leads to more stable behavior in the erode/dilate
evaluation when dragging the GM around in viewer, it is definitely
a different output and triggering the assert failures in Flutter.

When the kSampledRepeatedly behavior was added, there were some
regressions in performance on some platforms (higher end where math
is probably cheaper than another image). The current check for
"any" deferred color filter is pretty aggressive. Something like
an alpha multiply is probably always worth deferring. It also
happens to avoid the aliasing issue that tickles the Flutter test.

Bug: b/40042615
Bug: b/40040587
Change-Id: I7f89bc47ac8c3c0ea4b53b8f4036f6a166e58fa7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/796876
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>