Implement Myers' Sweep-Line algorithm
Implement a less efficient sweep line method from the
Myers' paper "AN O(ElogE+i) EXPECTED TIME ALGORITHM FOR
THE PLANAR SEGMENT INTERSECTION PROBLEM". This
implementation becomes faster than the brute force
method when the number of segments is > 50. As
the number of segments goes up, this Myers
implementation can be 100x faster than the brute force
method.

It uses exact arithmetic, and does not round. Round
is left for later stages.

TESTING:
This algorithm is compared with the brute force method
to make sure it generates the same answers over
randomly generated segments.

The paper is here:
https://drive.google.com/file/d/1lYw5m8ryC5C1DiG1NIHOV5tp4U4GdgH2/view?usp=drive_link

To see how this change fits in to future CLs consult:
https://skia-review.googlesource.com/c/skia/+/784551

Bug: b/297220808
Change-Id: Iab210810ec9c6e93c9a0cf2cfad9a50aa65b67bc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/798716
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Herb Derby <herb@google.com>