Allow MSVC "host" builds
Some targets (like libgrapheme) require compiling a build tool for the
host platform, running this tool on the host as part of the build, and
then using the output of the tool. The msvc toolchain did not specify a
separate host toolchain so everything was built for the target. This
leads to build time errors when using msvc to cross compile from x64 to
arm64.

This change introduces a host toolchain by following the example of the
gcc_like toolchain. In addition to templating the toolchain a number of
assignments are updated to use current_cpu and current_os instead of
target_cpu and target_os. This change does not yet make it possible to
set the msvc host to anything other than x64 as it is currently
hardcoded, but this may be done in the future.

This change allows libgrapheme to be cross compiled with msvc from x64
to arm64. Unfortunately, it is not yet possible to enable building it
on all the builders, as some builders do not have all the packages
needed to run an xSAN executable.

Change-Id: I097a69baedb499c4873034b7d501928209b805bc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/799461
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>