Remove InterfaceBlock::typeOwner().
This turns out to be unused. I'm not sure, but my hunch is that it
was originally added as a safeguard to ensure the the SymbolTable
associated with the InterfaceBlock's type would not get deleted
ahead of time. Since InterfaceBlocks are always created at the
top level of a program, it seems like this should be safe either
way.

Change-Id: I37460414998f9767d6dbfbf49cf4e23994cfdd31
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/799937
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>