[graphite] Support the MSAA16 mode in VelloRenderer
Added a ComputeStep definition for the MSAA16 variant of the fine stage.
There are now two separate pipeline instantiations for the fine stage,
one for area AA and one for MSAA16.

The mask LUT gets computed once and stored for the lifetime of a
VelloRenderer. The payload gets uploaded to the GPU at render time in a
new buffer suballocation.

Currently the compute PathAtlas always selects analytic AA for mask
renders.

Bug: b/285420658

Change-Id: Idcd05f13cc6a8bbda93b413bbd0c88d95c66a1f7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/801376
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>