Fix for-loop variable scoping.
For loop induction variables now share a symbol table with the
for-loop body statement, rather than having a symbol table that is
solely responsible for induction variables.

This approach is modeled after the GLSL grammar, which uses a
special `statement_no_new_scope` node for the body of a for loop.

Bug: b/320478438
Change-Id: Icf0beffa62386c74a269cff3505f0d1d0af84123
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/802796
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>