[graphite] Fix and optimize Transform scale factors and inverse
This started as a refactor to move the hidden `local_aa_radius`
function out of `AnalyticRRectRenderStep` and
`PerEdgeAAQuadRenderStep` into `Transform`. Besides removing duplicate
code, this would allow `CoverageMaskRenderStep` to use it for the AA
bounds outset needed by the ClipStack. To simplify that, the Renderer
and RenderStep now just have a flag saying the shape bounds needs to
add AA outsetting, and the virtual function is removed.

Once the local AA radius function was on Transform, the old X and Y
axis scale factors that were cached were replaced with the proper
min/max singular values of the transform. For non-perspective
matrices, the scale factors are now calculated once per transform and
cached across all draws that may use that.

Additionally, I optimized the scale factor calculation to only
compute the SVD when it was an affine matrix. For rect-stays-rect and
simpler, the scale factors are determined with absolute value. Since
I was now doing more transform-type-specific calculations in the
Transform constructor, it also felt reasonable to optimize the inverse
for the non-perspective cases.

However, while going through those optimizations and improvements for
scale factor and inverse, I realized that a lot of the old logic that
was in the concat() functions that tried to infer the type and scale
factors from the two input functions were invalid. Certainly now with
using singular values, the product of two matrices can have SVs
less than the product of the SVs of the input matrices. But it's also
possible for the type to reduce (whereas the old code always thought
it would be the maximum type of the two transforms), for instance if
you concat the inverse it reduces to the identity type.

Now that the inverse and scale factors were better optimized, I
figure that things are much simpler and consistent it all of the
concat operations just operate by calculating the matrix and re-doing
the type/scale factor extraction.


Bug: b/309704495
Change-Id: Ia8b408a2f91f4065a299a684585c91d2b3955754
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/802456
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>