[pdf] Allow setting the document language
PDF provides for a /Lang key in the catalog dictionary to explicitly
state the natural language of the document. This is of particular use to
screen readers when selecting a voice.

Add SkPDF::Metadata::fLang so the user can explicitly state the language
of the document. In the event this is not set, use the language of the
root of the structure element tree (when provided).

Bug: chromium:1362536
Change-Id: Ic30fb56815073e6b52d95b03bde306e46850452d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/804397
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>