Properly track `isBuiltin` on struct and array types.
Previously, we would assume that no structs or arrays were built
in, because in practice we didn't have any in our modules. Now
we will properly copy the state of ProgramConfig::fIsBuiltinCode
(which tracks whether or not are we building a module right now)
into struct and array types when they are created.

Bug: b/320462028
Change-Id: Ie1873303f55750a5dfac10ac1fecab3feae68ddf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/805517
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>