Fix fuzzer-discovered error with dead-variable elimination.
The fuzzer constructed a case where the inliner would inline
a struct constructor, and then dead-code analysis would eliminate
it. This caused the program usage analysis to become unbalanced.

Previously, we were assuming that for a BinaryExpression,
calling `usage.remove` on one of its sides was a safe way to
remove that side of the BinaryExpression from the program. This
does not work properly now that we are tracking struct usage,
because a BinaryExpression for a given type surprisingly
references the given type _three_ times:
- the entire BinaryExpression is that type
- the expression's left side is that type
- the expression's right side is that type

I also made sure that a StructDefinition tracks the usage of
nested structs properly.

Bug: oss-fuzz:66207
Change-Id: I2226661c74bef68982a628a6dab249ea19262fac
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/806696
Commit-Queue: Robert Phillips <robertphillips@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>