Fix fuzzer-discovered timeout by caching fIsAllowedInES2.
The fuzzer was able to synthesize a tree of very-deeply-nested
structs in order to force repeated scanning of the hierarchy,
eventually causing a timeout. As with other fields in StructType,
this can be fixed just by caching the result.

Bug: oss-fuzz:66263
Change-Id: Ib0fe64c88b12855b04271d2e4c154f7dec8f9b9b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/806776
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>