Match canvas2d sweep angle calculations in MotionMark slide, consolidate functions
Without the changes to canonicalize the starting and ending angles
that is handled at canvas2d, and not in drawArc(), the sweep angles
gradually increase as the slide animates. The paths end up then
wrapping around themselves over and over again and increases the
number of verbs that have to be chopped and uploaded. This was showing
as a steady increase in frame time when viewing with 's'.

This also adds some default definitions to the Stage to handle the
common operations of looping over the objects and handling key presses.

Change-Id: Ic9e288639ff59a991205f3e28b38dbe7504cf1a2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/806996
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>