Add Bazel target for modular build of SkResources
This updates the documentation of SkResources that clients will
be responsible for registering their codecs, not depending on what
Skia was compiled with.

It also adds an API that takes in an SkCodec in the cases where
callers can decode the image (e.g. via a known codec).

Change-Id: Ic0f63f746f45c8f0eacdab689883784fe7a33a31
Bug: b/40045064
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/807957
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>