[unicode] Fix SkBreakIterator_libgrapheme
SkBreakIterator_libgrapheme::isDone was comparing the line break index
with the total number of UTF-8 code units in the original input. Instead
compare against the number of line breaks. This allows fEnd to be
removed. fStart was always 0, so it too can be removed.

The inital state must be pointing to the beginning of the string, which
is added. There must also be a final "done" state at the end of the
string, which is added. Calling next must move the next break.

Change-Id: I0ed2683bcf09cde57377f83a2a7e6834b33cdc58
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/809076
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>