[skif] Consolidate bounds analysis, refine flags
This folds the decal-scale-factor analysis into analyzeBounds() and
adds additional checks to see if texels would be sampled outside of
the image subset. It separates whether or not the tiling effect is
"visible" and just whether or not the outside texels are hit from
the sampling filter mode.

The construction of the SkShader is now handled in a new helper
function that knows how to handle all the flag combinations except for
kRequiresLayerCrop, which must be handled externally (either by
adding a clipRect or by resolving the image beforehand). With the new
sampling aware flags, this helper function and the draw() functions
are able to switch to the fast constraint.

While the fast path can't be taken in a lot of cases yet since the
intermediate images are still created with tight bounds, some tests
were still optimizable. The FilterResultTest logic is updated to
detect when an action used an optimized shader based on the
skif::Stats. It then adds an extra validation step of resolving the
FilterResult with shader that uses no optimizations and requires them
to be identical.


Change-Id: I892d65c6fd2b1fd37c4205fa9550a14700670efe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/805316
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>