[graphite] Allow different mask Renderers for PathAtlas.
Some PathAtlas types would like use a different Renderer depending on
the data being stored in the atlas. For example, the SmallPathAtlas
will want to store both bitmap mask data and SDF data. To handle this,
in the PathAtlas case we defer determining the Renderer in
Device::chooseRenderer() until we add the Shape to the atlas.

Change-Id: I129a0b32088f0972d75783cf32faa5a6d4dcfd8e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/810997
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>