[skif] Move FilterResult subsetting to member function
The new subset() function, and the existing resolve() and AutoSurface
snap() functions are now updated to return FilterResult directly
instead of a pair of image and offset. This will make it easier for
subset to participate in pixel boundary tracking and to preserve known
pixel boundaries throughout those operations.

Bug: b/323886079
Change-Id: I194c2ab915bad3b9608b13489ca48cf59f91e762
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/812237
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>