[graphite] Make atlas-related recordUploads() calls have same signature.
In looking at adding SmallPathAtlas I saw that the text atlas and the
RasterPathAtlas have different upload signatures. The former uses an
UploadList and the latter uses the DrawContext. Using the UploadList
makes the calling path a little strange so I changed it to match
RasterPathAtlas.

Change-Id: I9999eda4339b1309c3c0770dd3f6b179cf1a0497
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/813036
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>