[ganesh] Actually let SkSpecialImage::asShader() skip subset clamping
skif::FilterResult was tracking whether or not the subset would be
necessary based on the backing store dimensions, but even when it
would pass in "strict=false" to asShader(), the Ganesh special image's
call to exactify() would skip updating the dimensions because the
proxy hadn't been instantiated yet.

Without being updated, the non-subset version of SkImageShader would
use the SkImage's dimensions, which had been updated to the backing
store. This meant that when creating the GrTextureEffect, it would
see that the subset was beyond the proxy's logical dimensions and
still add the clamping logic to the shader.

AFAICT, exactify() is only used by SkSpecialImage via the analysis
that skif::FilterResult has already done, so this modified version
of exactify() should be safe. It simply updates the dimensions to
match the backing store dimensions so that the secondary subset
check doesn't think clamping is required.

Bug: b/315351386
Change-Id: I6453b97cb2e89dc90eca65778ea92901697cb275
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/812516
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>