Remove MtlIsAppInBackground.
This was added as a guard for Flutter because it seemed like apps were
going into the background and continuing to submit to the command
buffer. However, the check only works if Skia is running in the main
thread, and this is not the case with Flutter. In addition this won't
work with app extensions. So depending on this is deceptive.

A better solution is for the app to control this because it knows via
events when it's entering/leaving background. So it would wait for
Skia's submission to finish when entering the background (via a
FinishCallback), and then not submit anything to Skia until the app
returns to foreground.

Change-Id: I6b277d0f1d375beb0c2f385b87cae1c76d731432
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/814796
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>