Allow sk_Caps in private runtime effects.
Note that SkRP does not actually implement the Setting expression
so SkRP will still fail to render them, but that's not a concern
yet. Going forward, if we want them to work with SkRP, we can add
them on an as-needed basis.

Change-Id: I102015d58625e706e8cc184cc1dc3e3c7c2a0c95
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/814797
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>