[bazel] GaneshGLSurfaceManager.cpp: Rewrite makeGLESSurfaceManager() to be more like makeVulkanSurfaceManager().
In https://skia-review.googlesource.com/c/skia/+/811033 I implemented GaneshVulkanSurfaceManager.cpp, which required defining a makeVulkanSurfaceManager() function. In that CL, I argued that it's better for the make***SurfaceManager() functions to take all surface parameters as inputs, even if some of those parameters are always the same for all configs. The reasoning is that it makes it easier to quickly see what "ingredients" go into making a surface.

This CL does the same for makeGLESSurfaceManager().

Bug: b/40045301
Change-Id: I33e6e267599302aa9dd98e8bc56b379e3ceaca7a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/811244
Commit-Queue: Leandro Lovisolo <lovisolo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>