[skif] Don't restrict layer bounds by desired output for deferred actions
Previously, the desired output was always folding into the layer bounds
but this actually confused later calls to analyzeBounds() into thinking
that a layer crop would be needed when it wasn't actually. This was
particularly common when mixing matrix transform image filters with an
SkCanvas that had a clipped dst.

Originally, restricting bounds tracking to the desired output was
useful while migrating image filters to FilterResult. Now that
everything is on the new system, they already factor the desired
output when creating their offscreen surfaces and these extra
intersections are unnecessary.

I updated `imageAndOffset` to now apply the desired output, which
required cleaning up some handling in rescale() and blur(), which
are lagging behind in the boundary tracking update. I also changed
the filter mode in the imagefilter_composed_transform GM to be lienar
so that we wouldn't trigger resolves purely for NN. Now the GM will
have no offscreens at all whether or not it's being drawn normally
or tiled for DDLs.

Bug: b/323885532
Change-Id: I8a32c2a548d08f75c95ce256a3e96534dd1917b5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/815476
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>