Adjust warnings for GCC 13
The AutoTMallocSelfMove test explicitly does a self-move as a test.
Adjust the warning pragmas it uses for Clang to also work with GCC.

The dangling-reference warning is essentially an attempt at a borrow
checker for C++. It is more of a lint than a warning and only mostly
works because it hard codes exceptions for view types in std. We expect
any issues it could find to be reported on by
sanitize-address-use-after-return and sanitize-address-use-after-scope.

Change-Id: I43bf515b1a59827557e6332988b6c687f32e182b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/816556
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>