Remove 2 SkParagraph Asserts
The first assert can be false when the text is right-alignd and the
input width is infinity, in which case SkParagraph skips layout and
there will be no lines.

The second assert can be false when [startIndex, endIndex) includes
a zero-sized placeholder. Unfortunately it's not as easy as changing
the implementaton of `TextLine::iterateThroughSingleRunByStyles`,
since the paragraph builder always adds a zero-sized placeholder at the
end of the text, so we need to distinguish a user added placeholder and
the EOT placeholder.

Change-Id: I00709cbf8048c6a59cf062527cb61a10f6ca39b7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/816356
Auto-Submit: Weiyu Huang <weiyuhuang@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Weiyu Huang <weiyuhuang@google.com>