Remove extra SK_APIs and set SKIA_IMPLEMENTATION in buffet build
SK_API should only be set in headers - it doesn't really do anything
in .cpp files.

For all our libraries, we should be setting SKIA_IMPLEMENTATION=1
when compiling it, but *not* when clients are using the exported
headers. To make this easy, I added it to the skia_cc_library
macro instead of having to remember it for every target.

Along the way, I realized we were setting defines in src/opts
and I'm pretty sure that is *not* what we want to do, so I fixed
that as well.

Bug: b/40045064
Change-Id: I94ad4d0024c70f6a2a6d18b038af0de95600ace1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/816557
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>