Expose enum SkPerlinNoiseShaderType to src/opts.
In a followup, SkRasterPipeline_opts is going to need access to
the SkPerlinNoiseShaderType enum. In this CL, the enum is now
factored out into a separate header, and the Bazel work for exposing
the header to src/opts is implemented. We can't #include
SkPerlinNoiseShaderImpl.h from SkRasterPipeline_opts.h directly
because it has _many_ additional dependencies; we do not want to
leak all of those into src/opts.

Bug: b/40045243
Change-Id: I0829e7b88b95a337d22c26a77212698901e8fcec
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/817476
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>