Add Raster Pipeline fused negative multiply-add primitive.
All of our RP platforms which support fma today also support fused
negative multiply-add. FMA calculates `a*b + c`; FNMA calculates
`-a*b + c`. (Another way to frame it is `a + b*c` versus `a - b*c`,
but the multiplication occurs first, so it's mentally simpler to
place it on the left-hand side, at least for me.)

This can occasionally produce better code, and the compiler doesn't
seem to be smart enough to emit this op on its own, so it's worth
having the primitive in our toolbox. (Unfortunately, ARM doesn't
appear to have an `a*b - c` primitive; I was looking for that, but
this is the closest I could get.)

I've updated inverse_mat2 to use the new primitive in order to get
some test coverage. This will cause the `runtime_intrinsics_matrix`
slide to use `nmad` while rendering the 2x2 matrix inverse tile.

Change-Id: I5d5ed0116235569d2b47db44245eeba969932fe1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/818416
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>