[graphite] Break out rect management in DrawAtlas.
Adds a new method addRect() which will try to reserve a rectangular
region in the DrawAtlas, updating the Rectanizer but not adding
any data. This will allow the RasterPathAtlas to rasterize directly
to that region in the DrawAtlas instead of having to copy already
rasterized data via addToAtlas().

Bug: b/294378744
Change-Id: I793b835b94a5f252f003b1fb886c03c28e25b95f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/818896
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Herb Derby <herb@google.com>